refactor: move i18n to core
