+++ src/core/i18n/lang/en.jsonc
