+++ src/core/i18n/lang/es.jsonc
