--- src/core/index.ts
@@ -1,8 +1,0 @@
-import { getAsset, importAsset, importAssets } from './assets';
-
-/** Asset Repository module. */
-export const assets = {
-    getAsset,
-    importAsset,
-    importAssets,
-};
